requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-filter==1.1.0
django-filters==0.2.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.0
pkg-resources==0.0.0
pymongo==3.6.1
pytz==2018.4
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
